requirements.txt
colorama==0.4.6
python_bcrypt==0.3.2
